FEAT: migrate frontend to Next.js with SSR via OpenNext Cloudflare adapter
Replace the React 19 + Vite SPA with Next.js App Router for server-side
rendering, deployed to Cloudflare Workers via the OpenNext adapter.

Architecture:
- Next.js App Router as the main entry point (built by OpenNext)
- Existing Hono API mounted via catch-all route at app/api/[[...path]]
- All 26 backend endpoints preserved unchanged
- Frontend components migrated to components/ with 'use client' directives
- Separate tsconfig for backend (tsconfig.backend.json) and frontend

Key changes:
- app/ restructured from Vite SPA to Next.js App Router
- Components moved from app/src/components/ to components/
- Navbar now uses Next.js Link for navigation between / and /admin
- Removed Vite, @cloudflare/vite-plugin, @vitejs/plugin-react deps
- Added next, @opennextjs/cloudflare, @tailwindcss/postcss deps
- wrangler.jsonc.template updated: main=.open-next/worker.js
