--- app/.env.example
@@ -1,4 +1,0 @@
-# Optional backend hostname for API calls
-# If not set, API calls will be made to the same origin (relative URLs)
-# Example: VITE_OPTIONAL_BACKEND_URL=https://api.example.com
-VITE_OPTIONAL_BACKEND_URL=
