--- app/.prettierrc
@@ -1,1 +1,0 @@
-{ "printWidth": 140, "singleQuote": true, "semi": true, "useTabs": false }
