+++ app/admin/page.tsx
@@ -1,0 +1,5 @@
+import AdminPageView from '@/components/AdminPageView';
+
+export default function Page() {
+  return <AdminPageView />;
+}
