+++ app/layout.tsx
@@ -1,0 +1,15 @@
+import type { Metadata } from 'next';
+import './globals.css';
+
+export const metadata: Metadata = {
+  title: 'AWS AccessBridge',
+  description: 'Multi-AWS account access management dashboard',
+};
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <html lang="en">
+      <body className="bg-gray-900 min-h-screen text-white">{children}</body>
+    </html>
+  );
+}
