--- app/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@rexezuge/aws-access-bridge-app",
-  "version": "1.0.0",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "cf-typegen": "wrangler types",
-    "clean": "rm -rf dist/",
-    "prebuild": "npm run cf-typegen && npm run prettier && npm run lint",
-    "build": "tsc -b && vite build",
-    "prerelease": "npm run clean && npm install",
-    "release": "npm run build",
-    "prettier": "prettier --write .",
-    "lint": "eslint --ext .ts,.tsx --fix ."
-  },
-  "devDependencies": {
-    "@cloudflare/vite-plugin": "^1.30.x",
-    "@tailwindcss/vite": "^4.1.x",
-    "@types/node": "^25.0.x",
-    "@types/react": "^19.2.x",
-    "@types/react-dom": "^19.2.x",
-    "@vitejs/plugin-react": "^5.1.x",
-    "eslint": "^9.39.x",
-    "eslint-plugin-react-hooks": "^7.0.x",
-    "eslint-plugin-react-refresh": "^0.4.x",
-    "globals": "^17.0.x",
-    "react": "^19.2.x",
-    "react-dom": "^19.2.x",
-    "terser": "^5.44.x",
-    "typescript": "^5.9.x",
-    "typescript-eslint": "^8.52.x",
-    "vite": "^7.3.x"
-  }
-}
