+++ app/page.tsx
@@ -1,0 +1,5 @@
+import HomePage from '@/components/HomePage';
+
+export default function Page() {
+  return <HomePage />;
+}
