--- app/src/vite-env.d.ts
@@ -1,9 +1,0 @@
-/// <reference types="vite/client" />
-
-interface ImportMetaEnv {
-  readonly VITE_OPTIONAL_BACKEND_URL?: string;
-}
-
-interface ImportMeta {
-  readonly env: ImportMetaEnv;
-}
