--- app/wrangler.jsonc
@@ -1,5 +1,0 @@
-{
-  "$schema": "node_modules/wrangler/config-schema.json",
-  "name": "aws-access-bridge-app",
-  "compatibility_date": "2025-11-20",
-}
