+++ next.config.ts
@@ -1,0 +1,8 @@
+import type { NextConfig } from 'next';
+import { initOpenNextCloudflareForDev } from '@opennextjs/cloudflare';
+
+initOpenNextCloudflareForDev();
+
+const nextConfig: NextConfig = {};
+
+export default nextConfig;
