+++ package.json
@@ -1,36 +1,45 @@
 {
   "name": "@rexezuge/aws-access-bridge",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
-    "cf-typegen": "wrangler types",
-    "buildApp": "cd app/ && npm run release",
+    "dev": "next dev",
+    "build": "opennextjs-cloudflare build",
+    "deploy": "opennextjs-cloudflare build && opennextjs-cloudflare deploy",
+    "deploy:fast": "opennextjs-cloudflare deploy",
+    "preview": "opennextjs-cloudflare build && opennextjs-cloudflare preview",
+    "cf-typegen": "wrangler types --env-interface CloudflareEnv ./cloudflare-env.d.ts",
     "test": "vitest run",
-    "predeploy": "npm install && npm run cf-typegen && npm run prettier && npm run lint && npm run tsc && npm test",
-    "deploy": "npm run buildApp && wrangler deploy --dry-run",
-    "postdeploy": "wrangler deploy",
-    "deploy:fast": "wrangler deploy",
-    "prettier": "prettier --write '!app' .",
-    "lint": "eslint --ext .ts,.tsx --fix --ignore-pattern app/ .",
-    "tsc": "tsc --noemit",
-    "migrations:apply": "yes | wrangler d1 migrations apply --remote aws-access-bridge-db"
+    "prettier": "prettier --write .",
+    "lint": "eslint --ext .ts,.tsx --fix .",
+    "tsc": "tsc --noEmit && tsc --noEmit --project tsconfig.backend.json"
+  },
+  "dependencies": {
+    "@opennextjs/cloudflare": "^1.19.x",
+    "aws4fetch": "^1.0.x",
+    "chanfana": "^3.3.x",
+    "hono": "^4.12.x",
+    "jose": "^6.2.x",
+    "next": "^15.3.x",
+    "react": "^19.2.x",
+    "react-dom": "^19.2.x"
   },
   "devDependencies": {
+    "@cloudflare/workers-types": "^4.20260414.1",
+    "@tailwindcss/postcss": "^4.1.x",
+    "@types/node": "^22.15.x",
+    "@types/react": "^19.2.x",
+    "@types/react-dom": "^19.2.x",
     "eslint": "^10.2.x",
     "eslint-config-prettier": "^10.1.x",
     "eslint-plugin-prettier": "^5.5.x",
     "globals": "^17.5.x",
     "jsonc-parser": "^3.3.x",
     "prettier": "^3.8.x",
-    "typescript": "^6.0.x",
+    "tailwindcss": "^4.1.x",
+    "typescript": "^5.8.x",
     "typescript-eslint": "^8.58.x",
     "vitest": "^4.1.4",
     "wrangler": "^4.82.x"
-  },
-  "dependencies": {
-    "aws4fetch": "^1.0.x",
-    "chanfana": "^3.3.x",
-    "hono": "^4.12.x",
-    "jose": "^6.2.x"
   }
 }
