+++ tsconfig.json
@@ -1,47 +1,39 @@
 {
   "compilerOptions": {
-    /* Visit https://aka.ms/tsconfig.json to read more about this file */
-
-    /* Set the JavaScript language version for emitted JavaScript and include compatible library declarations. */
-    "target": "es2021",
-    /* Specify a set of bundled library declaration files that describe the target runtime environment. */
-    "lib": ["es2021"],
-    /* Specify what JSX code is generated. */
-    "jsx": "react-jsx",
-
-    /* Specify what module code is generated. */
-    "module": "es2022",
-    /* Specify how TypeScript looks up a file from a given module specifier. */
-    "moduleResolution": "Bundler",
-    /* Specify type package names to be included without being referenced in a source file. */
-    "types": ["./worker-configuration.d.ts"],
-    /* Enable importing .json files */
-    "resolveJsonModule": true,
-
-    /* Allow JavaScript files to be a part of your program. Use the `checkJS` option to get errors from these files. */
+    "target": "ES2021",
+    "lib": ["dom", "dom.iterable", "ES2021"],
     "allowJs": false,
-    /* Enable error reporting in type-checked JavaScript files. */
     "checkJs": false,
-
-    /* Disable emitting files from a compilation. */
-    "noEmit": true,
-
-    /* Ensure that each file can be safely transpiled without relying on other imports. */
-    "isolatedModules": true,
-    /* Allow 'import x from y' when a module doesn't have a default export. */
-    "allowSyntheticDefaultImports": true,
-    /* Ensure that casing is correct in imports. */
-    "forceConsistentCasingInFileNames": true,
-
-    /* Enable all strict type-checking options. */
-    "strict": true,
-
-    /* Skip type checking all .d.ts files. */
     "skipLibCheck": true,
+    "strict": true,
+    "noEmit": true,
+    "incremental": true,
+    "module": "esnext",
+    "moduleResolution": "bundler",
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "jsx": "preserve",
+    "allowSyntheticDefaultImports": true,
+    "forceConsistentCasingInFileNames": true,
+    "esModuleInterop": true,
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ],
     "paths": {
-      "@/*": ["./src/*"]
-    }
+      "@/*": ["./src/*"],
+      "@/components/*": ["./components/*"]
+    },
+    "types": ["./worker-configuration.d.ts", "@cloudflare/workers-types"]
   },
-  "exclude": ["test", "app"],
-  "include": ["worker-configuration.d.ts", "src/**/*.ts"]
+  "include": [
+    "next-env.d.ts",
+    "cloudflare-env.d.ts",
+    "worker-configuration.d.ts",
+    "**/*.ts",
+    "**/*.tsx",
+    ".next/types/**/*.ts"
+  ],
+  "exclude": ["node_modules", "test", ".next", ".open-next"]
 }
